pyproject.toml
[project]
name = "toy_datasets"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Matt Clifford", email = "mc15445@bristol.ac.uk"},
]
dependencies = ["numpy>=2.3.5", "scipy>=1.16.3", "scikit-learn>=1.7.2", "pandas>=2.3.3", "ucimlrepo>=0.0.7", "tqdm>=4.67.1", "torch>=2.7.1", "torchvision>=0.22.1", "umap-learn>=0.5.9.post2", "matplotlib>=3.10.7"]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}


[project.optional-dependencies]
dev = [
    "pip>=25.3",
]
[tool.pdm]
distribution = false
